// src/dojo/generated/typescript/models.gen.ts
export interface BlisterPackItem {
    game_id: Number;
    idx: Number;
    blister_pack_id: Number;
    cost: Number;
    purchased: Boolean;
    
}